+++ src/Components/Autocomplete.php
@@ -6,7 +6,6 @@ namespace Hirtz\Location\Google\Components;
 
 use Hirtz\Location\Modules\Admin\Interfaces\AutocompleteInterface;
 use GuzzleHttp\RequestOptions;
-use Hirtz\Media\Helpers\Html;
 use yii\base\BaseObject;
 
 class Autocomplete extends BaseObject implements AutocompleteInterface
@@ -37,7 +36,7 @@ class Autocomplete extends BaseObject implements AutocompleteInterface
 
         foreach ($suggestions as $suggestion) {
             $results[] = [
-                'text' => Html::encode($suggestion['placePrediction']['text']['text']),
+                'text' => $suggestion['placePrediction']['text']['text'],
                 'value' => $suggestion['placePrediction']['placeId'],
             ];
         }